3) Servidor/requirements.txt
django
djangorestframework
djangorestframework-api-response==0.0.30
django-environ
django-cors-headers

channels
channels-redis
django-redis
Pillow
psycopg2

python-slugify

cryptography
argon2-cffi
argon2-cffi-bindings
djoser
PyJWT
djangorestframework-simplejwt

botocore
boto3
django-storages
whitenoise

uvicorn

paho-mqtt

Adafruit_DHT